Refac: Rename duplicator/migrator to copier
Copying is the right word:
- Migration implies the source is removed
- Duplication implies it's on the same server

So we rename to copy/copier everywhere.
