+++ backend/config/settings/base.py
@@ -257,7 +257,7 @@ LOGGING = {
             "propagate": True,
         },
         # httpx/httpcore log every request/connection at INFO/DEBUG, which generates
-        # a lot of noise (e.g. when running the workspace duplicator).
+        # a lot of noise (e.g. when running the workspace copier).
         "httpx": {
             "level": "WARNING",
             "propagate": True,
@@ -337,7 +337,7 @@ INSTALLED_APPS = [
     "oauth2_provider",
     "hexa.assistant",
     "hexa.mcp",
-    "hexa.workspace_duplicator",
+    "hexa.workspace_copier",
     "django_otp",
     "django_otp.plugins.otp_static",
     "django_otp.plugins.otp_email",
