+++ backend/hexa/workspace_copier/apps.py
@@ -1,0 +1,6 @@
+from hexa.app import CoreAppConfig
+
+
+class WorkspaceCopierConfig(CoreAppConfig):
+    name = "hexa.workspace_copier"
+    label = "workspace_copier"
