--- backend/hexa/workspace_duplicator/apps.py
@@ -1,6 +1,0 @@
-from hexa.app import CoreAppConfig
-
-
-class WorkspaceDuplicatorConfig(CoreAppConfig):
-    name = "hexa.workspace_duplicator"
-    label = "workspace_duplicator"
